requirements.txt
# PDFPilot Dependencies

# --- Core ---
fastapi==0.115.0
uvicorn[standard]==0.30.0
python-multipart==0.0.9
numpy==1.26.4
torch==2.4.1
sentence-transformers==3.0.1

# --- PDF Processing ---
PyPDF2==3.0.1
pdfplumber==0.11.4
pdf2image==1.17.0
python-docx==1.1.2
reportlab==4.1.0

# --- Image Processing ---
Pillow==10.4.0

# --- OCR (optional — requires Tesseract installed) ---
pytesseract==0.3.13
rapidocr-onnxruntime==1.2.3

# --- HTML to PDF (optional — requires wkhtmltopdf installed) ---
# pdfkit==1.0.0
